+++ src/main/java/com/example/cabonerfbe/request/UpdateUnitRequest.java
@@ -1,7 +1,6 @@
 package com.example.cabonerfbe.request;
 
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.AllArgsConstructor;
@@ -9,6 +8,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -20,7 +20,7 @@ public class UpdateUnitRequest {
     private String unitName;
 
     @Positive(message = "Conversion factor must be a positive number")
-    private double conversionFactor;
+    private BigDecimal conversionFactor;
 
     @NotNull(message = "isDefault must be specified")
     private Boolean isDefault;
