+++ src/main/java/com/example/cabonerfbe/response/MidpointSubstanceFactorsResponse.java
@@ -2,6 +2,7 @@ package com.example.cabonerfbe.response;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -18,11 +19,11 @@ public class MidpointSubstanceFactorsResponse {
     private String compartmentName;
     private String molecularFormula;
     private String alternativeFormula;
-    private Double individualist;
-    private Double hierarchist;
-    private Double egalitarian;
+    private BigDecimal individualist;
+    private BigDecimal hierarchist;
+    private BigDecimal egalitarian;
 
-    public void setMethodValue(String methodName, Double value) {
+    public void setMethodValue(String methodName, BigDecimal value) {
         switch (methodName) {
             case "Individualist":
                 this.individualist = value;
